requirements.txt
Flask
Flask-Mail
python-dotenv
gunicorn
Flask-SQLAlchemy
psycopg2-binary
